graminize/requirements.txt
# © 2024 Nokia
# Licensed under the BSD 3-Clause Clear License
# SPDX-License-Identifier: BSD-3-Clause-Clear

python-dotenv
click
jinja2
tomli>=1.1.0
tomli-w>=0.4.0
toml>=0.10
voluptuous